requirements.txt
python-dotenv
openai>=0.27.0
google-genai
requests
python-telegram-bot>=20.0
aiohttp>=3.9.0
web3>=6.0.0
eth-account>=0.10.0
eth-abi>=4.0.0
py-clob-client>=0.18.0
streamlit
pandas

